# stability/openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "StabilityAI REST API", "version": "v2beta" },
  "paths": {
    "/stable-image/generate/ultra": {
      "post": {
        "summary": "Text to image generation with Stable Image Ultra",
        "operationId": "generate_with_ultra_stable_image_generate_ultra_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ImageGenRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "x-ouro-output-asset-filter": "image",
        "x-ouro-output-asset-type": "file"
      }
    },
    "/stable-image/generate/core": {
      "post": {
        "summary": "Text to image generation with Stable Image Core",
        "operationId": "generate_with_core_stable_image_generate_core_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ImageGenRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "x-ouro-output-asset-filter": "image",
        "x-ouro-output-asset-type": "file"
      }
    },
    "/stable-image/generate/sd3": {
      "post": {
        "summary": "Text to image generation with Stable Diffusion 3.0",
        "operationId": "generate_with_sd3_stable_image_generate_sd3_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ImageGenRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "x-ouro-output-asset-filter": "image",
        "x-ouro-output-asset-type": "file"
      }
    },
    "/stable-image/control/sketch": {
      "post": {
        "summary": "Image-to-image generation for controlled variations of existing images or sketches",
        "operationId": "control_with_sketch_stable_image_control_sketch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ControlRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "x-ouro-output-asset-filter": "image",
        "x-ouro-output-asset-type": "file",
        "x-ouro-input-asset-filter": "image",
        "x-ouro-input-asset-type": "file"
      }
    },
    "/3d/stable-fast-3d": {
      "post": {
        "summary": "Generate 3D assets from a single 2D input image",
        "description": "Stable Fast 3D generates high-quality 3D assets from a single 2D input image.",
        "operationId": "fast_3d_3d_stable_fast_3d_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Fast3DRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "x-ouro-output-asset-filter": "3d",
        "x-ouro-output-asset-type": "file",
        "x-ouro-input-asset-filter": "image",
        "x-ouro-input-asset-type": "file"
      }
    },
    "/image-to-video": {
      "post": {
        "summary": "Generate a short video from an image",
        "description": "Generate a short video based on an initial image with Stable Video Diffusion, a latent video diffusion model.",
        "operationId": "image_to_video_image_to_video_post",
        "parameters": [
          {
            "name": "ouro-user-id",
            "in": "header",
            "required": false,
            "schema": { "type": "string", "title": "Ouro-User-Id" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ImageToVideoRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "x-ouro-output-asset-filter": "video",
        "x-ouro-output-asset-type": "file",
        "x-ouro-input-asset-filter": "image",
        "x-ouro-input-asset-type": "file"
      }
    }
  },
  "components": {
    "schemas": {
      "AspectRatioEnum": {
        "type": "string",
        "enum": [
          "1:1",
          "16:9",
          "21:9",
          "2:3",
          "3:2",
          "4:5",
          "5:4",
          "9:16",
          "9:21"
        ],
        "title": "AspectRatioEnum"
      },
      "ControlRequest": {
        "properties": {
          "file": { "$ref": "#/components/schemas/File" },
          "prompt": {
            "type": "string",
            "title": "Prompt",
            "description": "What you wish to see in the output image. A strong, descriptive prompt that clearly defines elements, colors, and subjects will lead to better results."
          },
          "control_strength": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Control Strength",
            "description": "How much influence, or control, the image has on the generation. Represented as a float between 0 and 1, where 0 is the least influence and 1 is the maximum.",
            "default": 0.7
          },
          "negative_prompt": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Negative Prompt",
            "description": "A blurb of text describing what you do not wish to see in the output image."
          }
        },
        "type": "object",
        "required": ["file", "prompt"],
        "title": "ControlRequest"
      },
      "Fast3DRequest": {
        "properties": {
          "file": { "$ref": "#/components/schemas/File" },
          "texture_resolution": {
            "$ref": "#/components/schemas/TextureResolutionEnum",
            "title": "Texture Resolution",
            "description": "Determines the resolution of the textures used for both the albedo (color) map and the normal map.",
            "default": 1024
          },
          "foreground_ratio": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Foreground Ratio",
            "description": "Controls the amount of padding around the object to be processed within the frame. A higher ratio means less padding and a larger object, while a lower ratio increases the padding.",
            "default": 0.85
          }
        },
        "type": "object",
        "required": ["file"],
        "title": "Fast3DRequest"
      },
      "File": {
        "properties": {
          "id": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Id"
          },
          "name": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Name"
          },
          "description": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Description"
          },
          "url": { "type": "string", "title": "Url" },
          "filename": { "type": "string", "title": "Filename" },
          "type": { "type": "string", "title": "Type" }
        },
        "type": "object",
        "required": ["url", "filename", "type"],
        "title": "File"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ImageGenRequest": {
        "properties": {
          "prompt": {
            "type": "string",
            "title": "Prompt",
            "description": "What you wish to see in the output image. A strong, descriptive prompt that clearly defines elements, colors, and subjects will lead to better results."
          },
          "negative_prompt": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Negative Prompt",
            "description": "A blurb of text describing what you do not wish to see in the output image."
          },
          "aspect_ratio": {
            "$ref": "#/components/schemas/AspectRatioEnum",
            "title": "Aspect Ratio",
            "description": "Controls the aspect ratio of the generated image.",
            "default": "1:1"
          }
        },
        "type": "object",
        "required": ["prompt"],
        "title": "ImageGenRequest"
      },
      "ImageToVideoRequest": {
        "properties": {
          "file": { "$ref": "#/components/schemas/File" },
          "seed": {
            "type": "integer",
            "title": "Seed",
            "description": "A seed value to use for the video generation.",
            "default": 42
          }
        },
        "type": "object",
        "required": ["file"],
        "title": "ImageToVideoRequest"
      },
      "TextureResolutionEnum": {
        "type": "string",
        "enum": ["1024", "2048", "512"],
        "title": "TextureResolutionEnum"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      }
    }
  }
}